requirements.txt
flask>=3.0
pillow>=10.0
gunicorn>=21.0
python-dotenv>=1.0
